move bookmark icon out off SSR cache code path
action bar is rendered inside the article and therefore ssr cached, but ssr render for cache is without react context – can't do gql queries and should also not render any data that is not busted by the ssr cache key
