+++ components/Feed/Feed.js
@@ -4,7 +4,7 @@ import { nest } from 'd3-collection'
 import { timeFormat } from '../../lib/utils/format'
 import Link from '../Link/Href'
 import withT from '../../lib/withT'
-import ActionBar from '../ActionBar/FeedActionBar'
+import ActionBar from '../ActionBar/Feed'
 import StickySection from './StickySection'
 import PropTypes from 'prop-types'
 import formatCredits from './formatCredits'
